Keep tests warning-free even with high warnings

+++ Sample/Tests/ViewControllerTests-PlainXCTest.m
@@ -42,7 +42,7 @@
     NSArray *touchUpActions = [sut.showAlertButton actionsForTarget:sut
                                                     forControlEvent:UIControlEventTouchUpInside];
 
-    XCTAssertEqual(touchUpActions.count, 1);
+    XCTAssertEqual(touchUpActions.count, (NSUInteger)1);
     XCTAssertEqualObjects(touchUpActions[0], @"showAlert:");
 }
 
@@ -53,7 +53,7 @@
     NSArray *touchUpActions = [sut.showActionSheetButton actionsForTarget:sut
                                                           forControlEvent:UIControlEventTouchUpInside];
 
-    XCTAssertEqual(touchUpActions.count, 1);
+    XCTAssertEqual(touchUpActions.count, (NSUInteger)1);
     XCTAssertEqualObjects(touchUpActions[0], @"showActionSheet:");
 }
 
@@ -71,7 +71,7 @@
     
     [sut showAlert:nil];
     
-    XCTAssertEqual(alertVerifier.presentedCount, 1);
+    XCTAssertEqual(alertVerifier.presentedCount, (NSUInteger)1);
 }
 
 - (void)testShowActionSheet_ShouldPresentModal
@@ -81,7 +81,7 @@
 
     [sut showActionSheet:nil];
 
-    XCTAssertEqual(alertVerifier.presentedCount, 1);
+    XCTAssertEqual(alertVerifier.presentedCount, (NSUInteger)1);
 }
 
 - (void)testShowAlert_ShouldPreferAlert
@@ -171,7 +171,7 @@
 
     [sut showAlert:nil];
 
-    XCTAssertEqual(alertVerifier.actionTitles.count, 3);
+    XCTAssertEqual(alertVerifier.actionTitles.count, (NSUInteger)3);
     XCTAssertEqualObjects(alertVerifier.actionTitles[0], @"Default");
     XCTAssertEqualObjects(alertVerifier.actionTitles[1], @"Cancel");
     XCTAssertEqualObjects(alertVerifier.actionTitles[2], @"Destroy");
@@ -184,7 +184,7 @@
 
     [sut showActionSheet:nil];
 
-    XCTAssertEqual(alertVerifier.actionTitles.count, 3);
+    XCTAssertEqual(alertVerifier.actionTitles.count, (NSUInteger)3);
     XCTAssertEqualObjects(alertVerifier.actionTitles[0], @"Default");
     XCTAssertEqualObjects(alertVerifier.actionTitles[1], @"Cancel");
     XCTAssertEqualObjects(alertVerifier.actionTitles[2], @"Destroy");
